backend : modify request method

+++ mobile/app/src/main/java/com/vixiloc/vcashiermobile/presentation/screens/employee/employees/components/ListItem.kt
@@ -55,38 +55,40 @@ fun EmployeeListItem(
                     style = MaterialTheme.typography.bodySmall.copy(fontSize = 10.sp)
                 )
             }
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                FilledIconButton(
-                    modifier = Modifier
-                        .size(32.dp)
-                        .background(
-                            color = MaterialTheme.colorScheme.primary.copy(alpha = 0.2f),
-                            shape = MaterialTheme.shapes.small
-                        ),
-                    shape = MaterialTheme.shapes.small,
-                    icon = Icons.Outlined.Key,
-                    onClick = {
-                        onReset(item)
-                    },
-                    iconSize = 16
-                )
-                FilledIconButton(
-                    modifier = Modifier
-                        .size(32.dp)
-                        .background(
-                            color = MaterialTheme.colorScheme.error.copy(alpha = 0.2f),
-                            shape = MaterialTheme.shapes.small
-                        ),
-                    shape = MaterialTheme.shapes.small,
-                    icon = Icons.Outlined.Delete,
-                    onClick = {
-                        onDelete(item)
-                    },
-                    iconTint = MaterialTheme.colorScheme.error,
-                    iconSize = 16
-                )
+            if(item.role != "admin") {
+                Row(
+                    horizontalArrangement = Arrangement.spacedBy(12.dp)
+                ) {
+                    FilledIconButton(
+                        modifier = Modifier
+                            .size(32.dp)
+                            .background(
+                                color = MaterialTheme.colorScheme.primary.copy(alpha = 0.2f),
+                                shape = MaterialTheme.shapes.small
+                            ),
+                        shape = MaterialTheme.shapes.small,
+                        icon = Icons.Outlined.Key,
+                        onClick = {
+                            onReset(item)
+                        },
+                        iconSize = 16
+                    )
+                    FilledIconButton(
+                        modifier = Modifier
+                            .size(32.dp)
+                            .background(
+                                color = MaterialTheme.colorScheme.error.copy(alpha = 0.2f),
+                                shape = MaterialTheme.shapes.small
+                            ),
+                        shape = MaterialTheme.shapes.small,
+                        icon = Icons.Outlined.Delete,
+                        onClick = {
+                            onDelete(item)
+                        },
+                        iconTint = MaterialTheme.colorScheme.error,
+                        iconSize = 16
+                    )
+                }
             }
         }
     }
